Implemented Session Configuration manager to keep reference to objects
