+++ Sources/RequestDL/Internals/Sources/Client/Client Manager/Models/Internals.ClientManager.Item.swift
@@ -1,0 +1,36 @@
+/*
+ See LICENSE for this package's licensing information.
+*/
+
+import Foundation
+
+extension Internals.ClientManager {
+
+    struct Item {
+        let configuration: Internals.Session.Configuration
+        let client: Internals.Client
+        let readAt: Date
+    }
+}
+
+extension Internals.ClientManager.Item {
+
+    static func createNew(
+        configuration: Internals.Session.Configuration,
+        client: Internals.Client
+    ) -> Internals.ClientManager.Item {
+        .init(
+            configuration: configuration,
+            client: client,
+            readAt: .init()
+        )
+    }
+
+    func updatingReadAt() -> Internals.ClientManager.Item {
+        .init(
+            configuration: configuration,
+            client: client,
+            readAt: .init()
+        )
+    }
+}
