+++ Sources/RequestDL/Internals/Sources/Client/Client Operation Queue/Models/QueueClientOperationDelegate.swift
@@ -1,0 +1,11 @@
+/*
+ See LICENSE for this package's licensing information.
+*/
+
+import Foundation
+
+@HTTPClientActor
+protocol QueueClientOperationDelegate: AnyObject {
+
+    func operationDidComplete(_ operation: Internals.ClientOperation)
+}
