+++ Sources/RequestDL/Internals/Sources/Client/HTTP Client Actor/HTTPClientActor.swift
@@ -1,0 +1,10 @@
+/*
+ See LICENSE for this package's licensing information.
+*/
+
+import Foundation
+
+@globalActor
+actor HTTPClientActor {
+    static let shared = HTTPClientActor()
+}
